REFACTOR

+++ lib/vtrc/common/data-queue.cpp
@@ -234,7 +234,7 @@ namespace vtrc { namespace common { namespace data_queue {
 
     namespace varint {
 
-        typedef vtrc::common::policies::varint_policy<uint32_t> policy_type;
+        typedef vtrc::common::policies::varint_policy<size_t> policy_type;
 
         queue_base *create_parser( size_t max_valid_length )
         {
@@ -260,7 +260,7 @@ namespace vtrc { namespace common { namespace data_queue {
 
     namespace fixint {
 
-        typedef vtrc::common::policies::fixint_policy<uint32_t> policy_type;
+        typedef vtrc::common::policies::fixint_policy<size_t> policy_type;
 
         queue_base *create_parser( size_t max_valid_length )
         {
